Arquitetura e Design de Projetos Java/09_ddd/escola/src/main/java/br/com/alura/escola/academico/aplicacao/aluno/matricula/MatricularAluno.java
package br.com.alura.escola.academico.aplicacao.aluno.matricula;

import br.com.alura.escola.academico.dominio.aluno.Aluno;
import br.com.alura.escola.academico.dominio.aluno.AlunoMatriculado;
import br.com.alura.escola.academico.dominio.aluno.RepositorioDeAlunos;
import br.com.alura.escola.shared.dominio.evento.PublicadorDeEventos;

public class MatricularAluno {
	// Use case
	// Padrão command
	private final RepositorioDeAlunos repositorio;
	private final PublicadorDeEventos publicador;

	public MatricularAluno(RepositorioDeAlunos repositorio, PublicadorDeEventos publicador) {
		this.repositorio = repositorio;
		this.publicador = publicador;
	}
	
	public void executar(MatricularAlunoDto dados) {
		Aluno novo = dados.criarAluno();
		
		this.repositorio.matricular(novo);
		
		AlunoMatriculado evento = new AlunoMatriculado(novo.getCpf());
		publicador.publicar(evento);
	}
}
